requirements.txt
dotenv==0.9.9
pillow==11.1.0
python-dotenv==1.0.1
screeninfo==0.8.1
